+++ tests/test_gifify_Sprite.py
@@ -1,13 +1,13 @@
 import json
 import pytest
-import gifify as gf
+from sprClasses import Sprite, SprCollection
 
 def test_create_Spr_img_file() :
-    spr: gf.Sprite = gf.Sprite.fromImageFile("exported_data/char_tm_img/tm201_01.png")
+    spr: Sprite = Sprite.fromImageFile("exported_data/char_tm_img/tm201_01.png")
     assert spr.img is not None
 
 def test_Spr_add_col() :
-    spr: gf.Sprite = gf.Sprite.fromImageFile("exported_data/char_tm_img/tm201_01.png")
+    spr: Sprite = Sprite.fromImageFile("exported_data/char_tm_img/tm201_01.png")
 
     jsonloc: str = "exported_data/char_tm_col/JSONs/tm201_01.json"
     spr.init_col_file(jsonloc)
@@ -26,5 +26,15 @@ def test_Spr_add_col() :
     assert len(spr.hitboxes) == len(exp_data["Hitboxes"])
 
 def test_Spr_weird_bg() :
-    spr: gf.Sprite = gf.Sprite.fromImageFile("exported_data/char_tm_img/tm030_06.png")
+    spr: Sprite = Sprite.fromImageFile("exported_data/char_tm_img/tm030_06.png")
     #spr.img.show()
+
+def test_SprColl_fromFiles() :
+    params = {
+        "names"             : 
+            ["tm201_0" + str(i) + ".png" for i in range(0,8)],
+        "img_path"          :
+            "exported_data/char_tm_img",
+        }
+
+    SprCollection.from_images_names()
